RAG-R1/requirements.txt
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
numpy
pandas
pybind11
ray
tensordict<0.6
transformers<4.48
vllm<=0.6.3
wandb
IPython
matplotlib
FlagEmbedding
flask
pdfplumber
faiss-gpu==1.7.3